Update PagesController.php
Missing controller was raising NotFoundHttpException. Added getTypes() and returned an empty array (not sure how to get the needed types).

Renamed getSources() to getTemplates(). getSources() was already defined in the parent controller.

The logic for editing a page is different than for content. I moved the edit logic to showForm() and referred to it from getEdit(). Same for getClone().